Move integration test in a separate directory, kafka is now using an ENV var to chose the version
