+++ src/integration/kotlin/com/github/imflog/schema/registry/security/BasicAuthenticationIT.kt
