+++ src/integration/kotlin/com/github/imflog/schema/registry/security/SslIT.kt
@@ -1,7 +1,7 @@
 package com.github.imflog.schema.registry.security
 
-import com.github.imflog.schema.registry.utils.Kafka5TestContainersUtils
 import com.github.imflog.schema.registry.tasks.config.ConfigTask
+import com.github.imflog.schema.registry.utils.KafkaTestContainersUtils
 import org.assertj.core.api.Assertions
 import org.gradle.internal.impldep.org.junit.rules.TemporaryFolder
 import org.gradle.testkit.runner.BuildResult
@@ -12,7 +12,7 @@ import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import java.io.File
 
-class SslIT : Kafka5TestContainersUtils() {
+class SslIT : KafkaTestContainersUtils() {
 
     private val folderRule: TemporaryFolder = TemporaryFolder()
     private lateinit var buildFile: File
