+++ src/integration/resources/secrets/registry.keystore.jks
