+++ src/integration/resources/secrets/registry.truststore.jks
